InputContainer 구현
